+++ Model/Skill.cs
@@ -14,7 +14,7 @@ namespace FirstAngularAPIAssignment.Model
         public string SkillName { get; set; }
 
         [Required(ErrorMessage = "Please Select Skill Experience.")]
-        public float SkillExperience { get; set; }
+        public int SkillExperience { get; set; }
         public int EmployeeId { get; set; }
         public virtual Employee Employee { get; set; }
